chore: downgrade ethers version and add jsbi as dependecy
