+++ package.json
@@ -83,7 +83,8 @@
     "@uniswap/sdk-core": "^7.7.2",
     "@uniswap/v3-sdk": "^3.25.2",
     "@uniswap/v4-sdk": "^1.21.4",
-    "ethers": "^6.0.0",
+    "ethers": "^5.7.2",
+    "jsbi": "^3.2.5",
     "viem": "^2.29.2",
     "wagmi": "^2.15.2"
   }
